% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_palettes.R
\name{pop_palette}
\alias{pop_palette}
\title{Color Palettes based on pop album covers}
\usage{
pop_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of palette. Select one:
\code{aqua}, \code{bangles}, \code{beck}, \code{boo},
\code{bwitched}, \code{deeelite},\code{hole}, \code{nodoubt},
\code{rickroll},\code{sclub7}, \code{spice},\code{steps}}

\item{n}{Number of colors desired.}

\item{type}{Either continuous or discrete.}
}
\value{
A vector of colors.
}
\description{
R package that contains color palettes based on pop album covers.
See also: https://github.com/johnmackintosh/metallicaRt for metallica palettes
}
\details{
and https://github.com/johnmackintosh/rockthemes for rock palettes
}
\examples{
pop_palette("aqua")

}
