% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_palette.R
\name{print_pal}
\alias{print_pal}
\title{print_pal}
\usage{
print_pal(x, ...)
}
\arguments{
\item{x}{palette name}

\item{...}{additional arguments}
}
\description{
print_pal
}
\examples{
\donttest{
print_pal('aqua')
}

}
